+++ apiserver/service/src/main/java/com/zipline/service/agentProperty/dto/response/AgentPropertyListResponseDTO.java
@@ -48,7 +48,7 @@ public class AgentPropertyListResponseDTO {
 		private Integer floor;
 		private Boolean hasElevator;
 		private Year constructionYear;
-		private Integer parkingCapacity;
+		private Double parkingCapacity;
 		private Double netArea;
 		private Double totalArea;
 		private String details;
